+++ apps/client/src/feature/auth/index.ts
@@ -1,5 +1,6 @@
 export { LocalSignUpForm } from '@/feature/auth/ui/signUp/LocalSignUpForm';
 export { implAuthService } from '@/feature/auth/service/authService';
+export type { AuthService } from '@/feature/auth/service/authService';
 export {
   AddGoogleSignUpModal,
   AddLocalSignUpModal,
