+++ apps/client/src/feature/company/index.tsx
@@ -1,0 +1,1 @@
+export { CreateCompanyForm } from '@/feature/company/ui/CreateCompanyForm';
