+++ apps/client/src/feature/post/index.tsx
@@ -1,8 +1,5 @@
-export {
-  implPostService,
-  type PostService,
-} from '@/feature/post/service/postService';
+export { implPostService } from '@/feature/post/service/postService';
+export type { PostService } from '@/feature/post/service/postService';
 
 export { PostDetailView } from '@/feature/post/ui/PostDetailView';
-export { CreateCompanyForm } from '@/feature/post/ui/CreateCompanyForm';
 export { CreatePostForm } from '@/feature/post/ui/CreatePostForm';
