+++ apps/client/src/feature/user/index.ts
@@ -1,2 +1,3 @@
 export { implUserService } from '@/feature/user/service/userService';
 export { MyInfo } from '@/feature/user/ui/MyInfo';
+export type { UserService } from '@/feature/user/service/userService';
