+++ apps/client/src/shared/context/ServiceContext.ts
@@ -1,10 +1,11 @@
 import { createContext } from 'react';
 
+import type { AuthService } from '@/feature/auth';
 import type { PostService } from '@/feature/post';
-import type { AuthService } from '@/feature/auth/service/authService';
+import type { UserService } from '@/feature/user';
 import type { EchoService } from '@/service/echoService';
 import type { ResumeService } from '@/service/resumeService';
-import type { UserService } from '@/feature/user/service/userService';
+import type { FileService } from '@/shared/file/fileService';
 
 export type ServiceContext = {
   echoService: EchoService;
@@ -12,6 +13,7 @@ export type ServiceContext = {
   postService: PostService;
   userService: UserService;
   resumeService: ResumeService;
+  fileService: FileService;
 };
 
 export const ServiceContext = createContext<ServiceContext | null>(null);
